Remove stale sockets in availableSockets (Fixes 408 timeout) (#6)
Applied a few fixes for socket that gives 408 timeout:

1. Server side should send FIN/ACK when client side sends FIN

When the client side closes a socket by using `socket.end()`, the server side should also close the socket, otherwise the `socket.on('end')` event won't be triggered.

Added these lines

```
    socket.once('end', () => {
      this.debug('socket ended');
      socket.end();
    });
```

2. Server side needs to resume the socket upon receiving it.

Added `socket.resume()`, otherwise the socket on the server side won't receive any event. To receive any event, there needs to be either a `socket.on('data', () => {})` listener, or equivalently, use `socket.resume()`. Something must be reading the socket otherwise the whole stream is paused, including the end signals.

3. Use latest socket first.

New sockets are pushed into the end of the array. And when accepting requests, we should use the new sockets from the end of the array. The old sockets at the start of the array may be in the process of expiring and they just haven't triggered the 'end' event yet/

```
const sock = this.availableSockets.pop();
```